HTTP transport (#411)
* Add config for http addresses

* HTTP transport WIP

* HTTP transport server cleanup

* HTTP Transport completion

* Either transport: add note of which connection we receive message from

* Transport cleanup

* refactor(p2p): Cleanup p2p code

* refactor(p2p): Cleanup p2p code

* refactor(clippy): Cleanup new clippy checks

* refactor(transport): More cleanup

+++ core/src/framing/capnp.rs
@@ -67,7 +67,7 @@ where
         let opts = capnp::message::ReaderOptions {
             // This remove security limit, but we keep reusing the reader and we eventually reach
             // that limit because of it.
-            traversal_limit_in_words: std::u64::MAX,
+            traversal_limit_in_words: std::usize::MAX as u64,
             ..Default::default()
         };
         let mut data = capnp_frame.exposed_data();
